third_party/blink/web_tests/platform/generic/printing/page-break-inside-avoid-expected.txt

Test for page-break-inside:avoid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1-1" is 1
PASS: page number of "page1-2" is 1
PASS: page number of "page2" is 2
PASS: page number of "page3" is 3
PASS: page number of "page4" is 4
PASS: page number of "page5" is 5
PASS: page number of "page8" is 8
PASS: page number of "page9-1" is 9
PASS: page number of "page9-2" is 9
PASS: page number of "page10" is 10
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
